data-processor/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
python-dotenv==1.0.0
httpx==0.25.1
nltk==3.8.1
spacy==3.7.2
beautifulsoup4==4.12.2 